#4-lab: in proces

+++ src/main/java/com/example/lab3/repositoriy/ShopRepository.java
@@ -5,8 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShopRepository extends CrudRepository<Shop, Long> {
-
+public interface ShopRepository extends CrudRepository<Shop, Integer> {
     @Override
     Iterable<Shop> findAll();
 }
